Log unexpected responses
This accomplishes a few long-standing TODO items, but also helps users
in showing exact registry error messages
